refactor: structure of project changed
