+++ .gitignore
@@ -3,7 +3,6 @@
 /node_modules
 
 # Logs
-logs
 *.log
 npm-debug.log*
 pnpm-debug.log*
