+++ src/routes/album/dto/update-album.dto.ts
