+++ src/routes/album/entities/album.entity.ts
@@ -1,5 +1,5 @@
-import ArtistEntity from 'src/artist/entities/artist.entity';
-import TrackEntity from 'src/track/entities/track.entity';
+import ArtistEntity from 'src/routes/artist/entities/artist.entity';
+import TrackEntity from 'src/routes/track/entities/track.entity';
 import {
   Column,
   Entity,
