+++ src/routes/artist/artist.controller.ts
