+++ src/routes/artist/artist.module.ts
@@ -1,8 +1,8 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AlbumModule } from 'src/album/album.module';
-import { FavsModule } from 'src/favs/favs.module';
-import { TrackModule } from 'src/track/track.module';
+import { AlbumModule } from 'src/routes/album/album.module';
+import { FavsModule } from 'src/routes/favs/favs.module';
+import { TrackModule } from 'src/routes/track/track.module';
 import { ArtistController } from './artist.controller';
 import { ArtistService } from './artist.service';
 import ArtistEntity from './entities/artist.entity';
