+++ src/routes/favs/models/favs.interface.ts
