+++ src/routes/track/track.module.ts
@@ -1,9 +1,9 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TrackController } from './track.controller';
 import { TrackService } from './track.service';
-import { ArtistModule } from 'src/artist/artist.module';
-import { AlbumModule } from 'src/album/album.module';
-import { FavsModule } from 'src/favs/favs.module';
+import { ArtistModule } from 'src/routes/artist/artist.module';
+import { AlbumModule } from 'src/routes/album/album.module';
+import { FavsModule } from 'src/routes/favs/favs.module';
 import TrackEntity from './entities/track.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
